backend/requirements_core.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
email-validator==2.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography>=41.0.0
httpx==0.25.2
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
aiolimiter==1.1.0
pydantic-settings==2.1.0
pyjwt==2.8.0
reportlab==4.0.9
pdfplumber==0.11.0
numpy>=1.26.4,<3.0.0
scipy>=1.16.1
scikit-learn>=1.3.2
psutil>=5.9.0
chromadb>=0.4.0
openai>=1.0.0
supabase>=2.0.0
postgrest>=0.13.0
